# swagger.yaml
openapi: 3.0.3
info:
  title: iSchool
  description: >-
    # A simple Student Management System
  version: 1.0.0
  contact: { email: ammuftau74@gmail.com, whatapp: +2340935095173 }
servers:
  - url: https://jigipay.onrender.com/api/v1
    description: This is the original api URL
    schemes: [https, http]
  - url: http://localhost:8080/api/v1
    description: THis is my local Server
paths:
...
  /users/login:
    post:
      tags:
        - 🔐 Authentication
      summary: Login user
      description: >-
        The API endpoint allows users to log into their accounts.
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user.email@domain.com
                password:
                  type: string
                  example: test@123
            example:
              email: user.email@domain.com
              password: test@123
      responses:
        '201':
          description: Login user
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '500'
            Date:
              schema:
                type: string
                example: Fri, 30 Jun 2023 06:44:09 GMT
            ETag:
              schema:
                type: string
                example: W/"1f4-1q38Sbu25p0UVmeYQmrEv2VrgPg"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            RateLimit-Limit:
              schema:
                type: string
                example: '500'
            RateLimit-Remaining:
              schema:
                type: string
                example: '495'
            RateLimit-Reset:
              schema:
                type: string
                example: '772'
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      __v:
                        type: number
                        example: 0
                      _id:
                        type: string
                        example: 649e795264e7dba29b726936
                      avatar:
                        type: object
                        properties:
                          _id:
                            type: string
                            example: 649e795264e7dba29b726935
                          url:
                            type: string
                            example: https://via.placeholder.com/200x200.png
                      createdAt:
                        type: string
                        example: '2023-06-30T06:42:26.163Z'
                      email:
                        type: string
                        example: user.email@domain.com
                      firstName:
                        type: string
                        example: john
                      lastName:
                        type: string
                        example: doe
                      phone:
                        type: string
                        example: '+23491234567'
                      isVerified:
                        type: boolean
                        example: false
                      loginType:
                        type: string
                        example: EMAIL_PASSWORD
                      role:
                        type: string
                        example: ADMIN
                      updatedAt:
                        type: string
                        example: '2023-06-30T06:42:26.247Z'
                      username:
                        type: string
                        example: doejohn
                      accessToken:
                        type: string
                        example: >-
                          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDllNzk1MjY0ZTdkYmEyOWI3MjY5MzYiLCJlbWFpbCI6IndhamUuc2h1YmhhbTExMUBnbWFpbC5jb20iLCJ1c2VybmFtZSI6IndhamVzaHViaGFtIiwicm9sZSI6IkFETUlOIiwiaWF0IjoxNjg4MTA3NDY5LCJleHAiOjE2ODgxOTM4Njl9.40qfVrWi8x8fGBRB9JBVCVsdMYvlwzjIRpffz5z7DK4
                      refreshToken:
                        type: string
                        example: >-
                          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDllNzk1MjY0ZTdkYmEyOWI3MjY5MzYiLCJpYXQiOjE2ODgxMDc0NjksImV4cCI6MTY4ODk3MTQ2OX0.7or55Coj0K_UY4YpzYzkNR9hcOIhusxk9FRz7eMV8Ho
                  message:
                    type: string
                    example: >-
                      User log in successfully.
                  success:
                    type: boolean
                    example: true
              examples:
                Login user:
                  value:
                    user:
                      __v: 0
                      _id: 649e795264e7dba29b726936
                      avatar:
                        publid_id: 649e795264e7dba29b726935
                        url: https://via.placeholder.com/200x200.png
                      createdAt: '2023-06-30T06:42:26.163Z'
                      firstame: john
                      lastName: djohn
                      phone: '+2345600909'
                      email: user.email@domain.com
                      username: doejohn
                      isVerified: false
                      loginType: EMAIL_PASSWORD
                      role: User
                      accessToken:
                        example: >-
                          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDllNzk1MjY0ZTdkYmEyOWI3MjY5MzYiLCJlbWFpbCI6IndhamUuc2h1YmhhbTExMUBnbWFpbC5jb20iLCJ1c2VybmFtZSI6IndhamVzaHViaGFtIiwicm9sZSI6IkFETUlOIiwiaWF0IjoxNjg4MTA3NDY5LCJleHAiOjE2ODgxOTM4Njl9.40qfVrWi8x8fGBRB9JBVCVsdMYvlwzjIRpffz5z7DK4
                      refreshToken:
                        example: >-
                          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDllNzk1MjY0ZTdkYmEyOWI3MjY5MzYiLCJpYXQiOjE2ODgxMDc0NjksImV4cCI6MTY4ODk3MTQ2OX0.7or55Coj0K_UY4YpzYzkNR9hcOIhusxk9FRz7eMV8Ho

                      updatedAt: '2023-06-30T06:42:26.247Z'
                    message: >-
                      User log in successfully.

                    success: true
...
  /users/assign-role/{userId}:
    post:
      tags:
        - 🔐 Authentication
      summary: Assign role
      description: >-
        The API endpoint allows administrators to change the roles of users
        within the system.


        This functionality is restricted to users with the `ADMIN` role,
        ensuring that only authorized personnel can modify user roles.


        It provides an essential tool for managing user permissions and ensuring
        proper authorization levels within the system.
      operationId: assignRole
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  example: ADMIN
            example:
              role: ADMIN
      responses:
        '200':
          description: Assign role
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '81'
            Date:
              schema:
                type: string
                example: Thu, 01 Jun 2023 18:21:25 GMT
            ETag:
              schema:
                type: string
                example: W/"51-BNIeEDSuAnaNdBbrTmV3XGTq/UI"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties: {}
                  message:
                    type: string
                    example: Role changed for the user
                  success:
                    type: boolean
                    example: true
              examples:
                Assign role:
                  value:
                    data: {}
                    message: Role changed for the user
                    success: true
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          example: 6489626dc265f0aa2dbb73a4
...